Don't try to delete non-existent file when no preprocessor.

+++ old_branches/omni4_0_develop/omni/src/tool/omniidl/python/omniidl/main.py
@@ -29,6 +29,9 @@
 
 # $Id$
 # $Log$
+# Revision 1.17.2.12  2002/10/28 12:20:45  dgrisby
+# Editing mishap.
+#
 # Revision 1.17.2.11  2002/10/28 11:56:50  dgrisby
 # Work around VC++ 7 problem with FILE* change.
 #
@@ -491,7 +494,7 @@ def main(argv=None):
             if cd_to is not None:
                 os.chdir(old_wd)
 
-            if temp_file:
+            if temp_file and not no_preprocessor:
                 os.remove(temp_file)
 
             idlast.clear()
